+++ packages/node-app/src/utils/model.utils.ts
