+++ packages/node-app/src/utils/validations.utils.ts
